Refactor product and customer models to use float for price and quantity; introduce TagProdotto enum for product categorization

+++ TheMerchant/ProdottiDisponibili.cs
@@ -1,5 +1,5 @@
 public static class Disponibili
 {
-    public static Dictionary<Prodotto, (int Quantita, int Prezzo)> Prodotti { get; set; } =
-     new Dictionary<Prodotto,(int Quantita, int Prezzo)>();
+    public static Dictionary<Prodotto, (int Quantita, float Prezzo)> Prodotti { get; set; } =
+     new Dictionary<Prodotto,(int Quantita, float Prezzo)>();
 }
